requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qrcode[pil]>=7.4.2
pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
beautifulsoup4==4.12.2
psutil==5.9.8
httpx==0.25.2
